added user CRUD and fixed some bugs

+++ boostan/view.py
@@ -1,6 +1,6 @@
 from flask import render_template, flash, redirect, url_for, session, request
 from boostan import app, db
-from boostan.forms import RegisterForm, LoginForm, SubjectInsertForm
+from boostan.forms import RegisterForm, LoginForm
 from MySQLdb import cursors
 import hashlib
 
@@ -61,15 +61,98 @@ def user_list():
     if not "loggedin" in session:
         return redirect(url_for("login_page"))
 
+    cur = db.connection.cursor(cursors.DictCursor)
+    cur.execute("SELECT * FROM user ORDER BY ID")
+    users = cur.fetchall()
+    cur.close()
+    return render_template("user_list.html", users=users)
+
 
 @app.route("/user_detail")
 def user_detail():
     if not "loggedin" in session:
         return redirect(url_for("login_page"))
 
+    id = request.args.get("id")
+    if id:
+        cur = db.connection.cursor(cursors.DictCursor)
+        cur.execute(f"SELECT * FROM user where ID={id}")
+        user = cur.fetchone()
+        cur.close()
+        if user:
+            return render_template("user_detail.html", id=id, user=user)
+        flash("this subject doesn't exist!", category="danger")
+        return redirect(url_for("user_list"))
+
+
+@app.route("/users/update", methods=["POST"])
+def user_update():
+    if not "loggedin" in session:
+        return redirect(url_for("login_page"))
+    if request.method == "POST":
+        id = request.form["id"]
+        first_name = request.form["first_name"]
+        last_name = request.form["last_name"]
+        username = request.form["username"]
+        email = request.form["email"]
+        cur = db.connection.cursor(cursors.DictCursor)
+        cur.execute(f"SELECT username, email FROM user WHERE id={id}")
+        current_user = cur.fetchone()
+        cur.execute(f"SELECT username FROM user WHERE username='{username}'")
+        edited_user = cur.fetchone()
+        cur.execute(f"SELECT email FROM user WHERE email='{email}'")
+        edited_email = cur.fetchone()
+        if edited_user:
+            if edited_user["username"] != current_user["username"]:
+                cur.close()
+                flash("this user already exist", category="danger")
+                return redirect(url_for("user_list"))
+
+        if edited_email:
+
+            if edited_email["email"] != current_user["email"]:
+                cur.close()
+                flash("this user already exist", category="danger")
+                return redirect(url_for("user_list"))
+
+        if session["ID"] == int(id) and request.form["password"]:
+            password = request.form["password"]
+            password = hashed_password(password)
+            cur.execute(
+                f"UPDATE user SET first_name='{first_name}', last_name='{last_name}', username ='{username}', email='{email}', password='{password}' WHERE id={id}"
+            )
+        else:
+            cur.execute(
+                f"UPDATE user SET first_name='{first_name}', last_name='{last_name}', username ='{username}', email='{email}' WHERE id={id}"
+            )
+        db.connection.commit()
+        cur.close()
+        flash("Data Updated Successfully", category="success")
+        return redirect(url_for("user_list"))
+
+
+@app.route("/users/delete", methods=["GET"])
+def user_delete():
+
+    if not "loggedin" in session:
+        return redirect(url_for("login_page"))
+    id = request.args.get("id")
+    cur = db.connection.cursor()
+    cur.execute(f"DELETE FROM user WHERE id={id}")
+
+    cur.execute("ALTER TABLE subject AUTO_INCREMENT = 1")
+    db.connection.commit()
+
+    flash("Record Has Been Deleted Successfully", category="success")
+    if session["ID"] == int(id):
+        return redirect(url_for("logout_page"))
+    return redirect(url_for("user_list"))
+
 
 @app.route("/subjects/insert", methods=["POST"])
 def subject_insert():
+    if not "loggedin" in session:
+        return redirect(url_for("login_page"))
     if request.method == "POST":
 
         name = request.form["name"]
@@ -78,15 +161,17 @@ def subject_insert():
         if cur.fetchone():
             cur.close()
             flash("this subject already exist", category="danger")
-            return redirect(url_for("admin_user_home"))
+            return redirect(url_for("subject_list"))
         cur.execute(f"INSERT INTO subject (name) VALUES ('{name}')")
         db.connection.commit()
         flash("Data Inserted Successfully", category="success")
-        return redirect(url_for("admin_user_home"))
+        return redirect(url_for("subject_list"))
 
 
 @app.route("/subjects/update", methods=["POST"])
 def subject_update():
+    if not "loggedin" in session:
+        return redirect(url_for("login_page"))
     if request.method == "POST":
         id = request.form["id"]
         name = request.form["name"]
@@ -95,23 +180,26 @@ def subject_update():
         if cur.fetchone():
             cur.close()
             flash("this subject already exist", category="danger")
-            return redirect(url_for("admin_user_home"))
+            return redirect(url_for("subject_list"))
 
         cur.execute(f"UPDATE subject SET name='{name}' WHERE id={id}")
         db.connection.commit()
         flash("Data Updated Successfully", category="success")
-        return redirect(url_for("admin_user_home"))
+        return redirect(url_for("subject_list"))
 
 
 @app.route("/subjects/delete", methods=["GET"])
 def subject_delete():
+    if not "loggedin" in session:
+        return redirect(url_for("login_page"))
     id = request.args.get("id")
     cur = db.connection.cursor()
     cur.execute(f"DELETE FROM subject WHERE id={id}")
-    db.connection.commit()
 
-    flash("Record Has Been Deleted Successfully")
-    return redirect(url_for("admin_user_home"))
+    cur.execute("ALTER TABLE subject AUTO_INCREMENT = 1")
+    db.connection.commit()
+    flash("Record Has Been Deleted Successfully", category="success")
+    return redirect(url_for("subject_list"))
 
 
 @app.route("/register", methods=["GET", "POST"])
@@ -128,12 +216,12 @@ def signup_page():
         db.connection.commit()
 
         cur.execute(f"SELECT ID FROM user WHERE username = '{form.username.data}';")
-        id = cur.fetchone()
+        user_id = cur.fetchone()
 
         cur.close()
 
         session["loggedin"] = True
-        session["ID"] = id
+        session["ID"] = user_id["ID"]
         session["username"] = form.username.data
 
         flash(
@@ -186,7 +274,7 @@ def login_page():
 def logout_page():
 
     session.pop("loggedin", None)
-    session.pop("id", None)
+    session.pop("ID", None)
     session.pop("username", None)
 
     return redirect(url_for("login_page"))
